+++ tests/unit/vnlib.plugins.extensions.loading/Configuration/ValidateTests.cs
@@ -1,6 +1,32 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+/*
+* Copyright (c) 2026 Vaughn Nugent
+* 
+* Library: VNLib
+* Package: VNLib.Plugins.Extensions.Loading.Tests
+* File: ValidateTests.cs 
+*
+* ValidateTests.cs is part of VNLib.Plugins.Extensions.Loading.Tests which is part of the larger 
+* VNLib collection of libraries and utilities.
+*
+* VNLib.Plugins.Extensions.Loading.Tests is free software: you can redistribute it and/or modify 
+* it under the terms of the GNU Affero General Public License as 
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* VNLib.Plugins.Extensions.Loading.Tests is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+* GNU Affero General Public License for more details.
+*
+* You should have received a copy of the GNU Affero General Public License
+* along with this program.  If not, see https://www.gnu.org/licenses/.
+*/
 
-namespace VNLib.Plugins.Extensions.Loading.Configuration.Tests
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+using VNLib.Plugins.Extensions.Loading.Configuration;
+
+namespace VNLib.Plugins.Extensions.Loading.Tests.Configuration
 {
     [TestClass()]
     public class ValidateTests
@@ -18,13 +44,13 @@ namespace VNLib.Plugins.Extensions.Loading.Configuration.Tests
         public void NotNullTest()
         {
             Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotNull<object>(null, nameof(NotNullTest)));
-            Validate.NotNull<object>(new object(), nameof(NotNullTest));
+            Validate.NotNull(new object(), nameof(NotNullTest));
 
             //Test strings
             Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotNull<string>(null, nameof(NotNullTest)));
             Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotNull("", nameof(NotNullTest)));
             Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotNull("        ", nameof(NotNullTest)));
-            Validate.NotNull<string>("Hello", nameof(NotNullTest));
+            Validate.NotNull("Hello", nameof(NotNullTest));
         }
 
         [TestMethod()]
@@ -39,8 +65,8 @@ namespace VNLib.Plugins.Extensions.Loading.Configuration.Tests
         {
             Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotEqual(5, 5, nameof(NotEqualTest)));
             // Test: Validate.NotEqual should handle null values
-            Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotEqual<string>(null!, "test", nameof(NotEqualTest)));
-            Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotEqual<string>("test", null!, nameof(NotEqualTest)));
+            Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotEqual(null!, "test", nameof(NotEqualTest)));
+            Assert.ThrowsExactly<ConfigurationValidationException>(() => Validate.NotEqual("test", null!, nameof(NotEqualTest)));
             
             Validate.NotEqual(5, 10, nameof(NotEqualTest));
             Validate.NotEqual("test1", "test2", nameof(NotEqualTest));